feat: OLED and RFID_Card * fix: change Pico by Nano * refactor: RFID_Card library NOTE: hardware/flash.h unsupported on Pico
