+++ Telecommande/include/connection.h
@@ -20,19 +20,26 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#ifndef RFID_Card_h
-#define RFID_Card_h
+#ifndef __CONNECTION_H
+#define __CONNECTION_H
 
-#include <Arduino.h>
+#if defined(__AVR_ATmega328__) || defined(__AVR_ATmega328P__)
+// Arduino NANO OR UNO
 
-class RFID_Card {
-  public:
-    boolean cardCheck(byte Code[4]);
-    boolean saveCard(byte Code[4]);
-    int cardNumber(void);
-    void clear(void);
-  private:
-    int nbr = 0;
-};
+#include <SoftwareSerial.h>
+
+#define RX 11
+#define TX 12
+
+SoftwareSerial Serial1(RX, TX);
+
+#elif defined(__AVR_ATmega2560__)
+// Arduino Mega 2560
+
+#else
+
+#error “Unsupported board selected!”
+
+#endif
 
 #endif
