+++ Telecommande/platformio.ini
@@ -8,11 +8,11 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
-[env:pico]
-platform = raspberrypi
-board = pico
+[env:nanoatmega328]
+platform = atmelavr
+board = nanoatmega328
 framework = arduino
 lib_deps = 
 	adafruit/Adafruit BusIO@^1.11.3
-	alexgyver/GyverOLED@^1.3.2
+	adafruit/Adafruit SSD1306@^2.5.3
 	powerbroker2/SerialTransfer@^3.1.2
